Use composition to access ModalDialog instead of inheritance
gnome-shell 3.34 made ModalDialog a GObject, which is difficult to
extend without breaking compatibility with the non-GObject version
in older versions of gnome-shell. Constructing a GObject looks the
same to JS code as constructing a non-GObject, so it should be
safe to use the same code in gnome-shell 3.32 and 3.34.

Supports gnome-shell 3.34.